rate limiter

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Http\Resources\PostResource;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +16,11 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->get());
+        $user = request()->user();
+        $posts = $user->posts()->paginate();
+        return PostResource::collection($posts);
+
+        //return PostResource::collection(Post::with('author')->get());
     }
 
     /**
@@ -24,7 +29,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id;
 
         $post = Post::create($data);
 
@@ -36,6 +41,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
         return response()->json($post);
     }
 
@@ -58,6 +64,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
         $post->delete();
         return response()->json(null, 204);
     }
